// app/commons/LoaderBouncing/types.ts
export type LoaderPosition = 'absolute' | 'relative';
export type BouncingDotSize = 'sm' | 'md' | 'lg';
export type BouncingDotColor =
  | 'basic'
  | 'social'
  | 'notification'
  | 'offers'
  | 'eprescriptions'
  | 'epayments'
  | 'error'
  | 'primary'
  | 'secondary'
  | 'default'
  | 'toolbarSubmenu'
  | 'infobar'
  | 'navbar'
  | 'sidebar'
  | 'defaultNeutral'
  | 'secondaryNeutral'
  | 'field'
  | 'onetPremium';